added deep search param  to parser
