+++ .env
@@ -1,2 +1,2 @@
 EXT=
-FILE_PATH=D:\Stuff\Projects
+FILE_PATH=
